boards UseGuards 적용 config import 및 적용 entity relation 적용

+++ src/boards/boards.service.ts
@@ -4,6 +4,7 @@ import {Board} from "./board.entity";
 import {Repository} from "typeorm";
 import {CreateBoardDto} from "./dto/create-board.dto";
 import {BoardStatus} from "./board-status.enum";
+import {User} from "../auth/user.entity";
 
 
 @Injectable()
@@ -15,17 +16,24 @@ export class BoardsService {
     ) {}
 
     async getAllBoards(): Promise<Board[]> {
-        return this.boardRepository.find();
+        return await this.boardRepository.find();
+    }
+    async getUserBoards(user: User): Promise<Board[]> {
+        console.log(user)
+        return await this.boardRepository.findBy({'userId': user.id});
     }
 
-    async createBoard(createBoardDto: CreateBoardDto): Promise<Board> {
+    async createBoard(
+        createBoardDto: CreateBoardDto,
+        user: User): Promise<Board> {
 
         const {title, description} = createBoardDto;
 
         const board = this.boardRepository.create({
             title,
             description,
-            status: BoardStatus.PUBLIC
+            status: BoardStatus.PUBLIC,
+            user
         });
 
         await this.boardRepository.save(board);
@@ -44,8 +52,8 @@ export class BoardsService {
     }
 
 
-    async deleteBoardById(id: number): Promise<Object> {
-        const result = await this.boardRepository.delete(id)
+    async deleteBoardById(id: number, user: User): Promise<Object> {
+        const result = await this.boardRepository.delete({id, 'userId': user.id})
         if (result.affected === 0) {
             throw new NotFoundException()
         } else {
